# Use no-op subsegments for no-op segments and don't generate IDs for n…