--- a/aws-xray-recorder-sdk-core/src/main/java/com/amazonaws/xray/AWSXRay.java
+++ b/aws-xray-recorder-sdk-core/src/main/java/com/amazonaws/xray/AWSXRay.java
@@ -102,9 +102,13 @@ public class AWSXRay {
         return globalRecorder.beginSegment(name, traceId, parentId);
     }
 
+    /**
+     * @deprecated Use {@code AWSXray.getGlobalRecorder().beginNoOpSegment() }.
+     */
+    @Deprecated
     @Nullable
     public static Segment beginDummySegment() {
-        return globalRecorder.beginDummySegment();
+        return globalRecorder.beginNoOpSegment();
     }
 
     public static void endSegment() {