--- a/aws-xray-recorder-sdk-core/src/main/java/com/amazonaws/xray/entities/Cause.java
+++ b/aws-xray-recorder-sdk-core/src/main/java/com/amazonaws/xray/entities/Cause.java
@@ -46,12 +46,10 @@ public class Cause {
     private final List<ThrowableDescription> exceptions;
 
     public Cause() {
-        //id = Entity.generateId();
         this(new ArrayList<>(), new ArrayList<>());
     }
 
     Cause(List<String> paths, List<ThrowableDescription> exceptions) {
-        //id = Entity.generateId();
         this.paths = paths;
         this.exceptions = exceptions;
     }