--- a/aws-xray-recorder-sdk-core/src/main/java/com/amazonaws/xray/entities/FacadeSegment.java
+++ b/aws-xray-recorder-sdk-core/src/main/java/com/amazonaws/xray/entities/FacadeSegment.java
@@ -69,6 +69,11 @@ public class FacadeSegment extends EntityImpl implements Segment {
         throw new UnsupportedOperationException(MUTATION_UNSUPPORTED_MESSAGE);
     }
 
+    @Override
+    public boolean isRecording() {
+        return sampled;
+    }
+
     /**
      * Unsupported as FacadeSegments cannot be mutated.
      * @throws UnsupportedOperationException in all cases