--- a/aws-xray-recorder-sdk-core/src/main/java/com/amazonaws/xray/entities/NoOpList.java
+++ b/aws-xray-recorder-sdk-core/src/main/java/com/amazonaws/xray/entities/NoOpList.java
@@ -1,0 +1,153 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License").
+ * You may not use this file except in compliance with the License.
+ * A copy of the License is located at
+ *
+ *  http://aws.amazon.com/apache2.0
+ *
+ * or in the "license" file accompanying this file. This file is distributed
+ * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ * express or implied. See the License for the specific language governing
+ * permissions and limitations under the License.
+ */
+
+package com.amazonaws.xray.entities;
+
+import java.util.Collection;
+import java.util.Collections;
+import java.util.Iterator;
+import java.util.List;
+import java.util.ListIterator;
+import java.util.Map;
+
+/**
+ * An empty {@link Map} which does nothing, but unlike {@link Collections#emptyList()} won't throw
+ * {@link UnsupportedOperationException}.
+ */
+// We don't actually use the type parameter so nullness annotations would be ignored anyways so no need to annotate.
+@SuppressWarnings("all")
+class NoOpList implements List<Object> {
+
+    static <T> List<T> get() {
+        return (List<T>) INSTANCE;
+    }
+
+    private static final NoOpList INSTANCE = new NoOpList();
+
+    private static final Object[] EMPTY_ARRAY = new Object[0];
+
+    @Override
+    public int size() {
+        return 0;
+    }
+
+    @Override
+    public boolean isEmpty() {
+        return true;
+    }
+
+    @Override
+    public boolean contains(Object o) {
+        return false;
+    }
+
+    @Override
+    public Iterator<Object> iterator() {
+        return Collections.emptyIterator();
+    }
+
+    @Override
+    public Object[] toArray() {
+        return EMPTY_ARRAY;
+    }
+
+    @Override
+    public <T> T[] toArray(T[] a) {
+        return a;
+    }
+
+    @Override
+    public boolean add(Object o) {
+        return true;
+    }
+
+    @Override
+    public boolean remove(Object o) {
+        return false;
+    }
+
+    @Override
+    public boolean containsAll(Collection<?> c) {
+        return false;
+    }
+
+    @Override
+    public boolean addAll(Collection<?> c) {
+        return false;
+    }
+
+    @Override
+    public boolean addAll(int index, Collection<?> c) {
+        return false;
+    }
+
+    @Override
+    public boolean removeAll(Collection<?> c) {
+        return false;
+    }
+
+    @Override
+    public boolean retainAll(Collection<?> c) {
+        return false;
+    }
+
+    @Override
+    public void clear() {
+    }
+
+    @Override
+    public Object get(int index) {
+        return null;
+    }
+
+    @Override
+    public Object set(int index, Object element) {
+        return null;
+    }
+
+    @Override
+    public void add(int index, Object element) {
+    }
+
+    @Override
+    public Object remove(int index) {
+        return null;
+    }
+
+    @Override
+    public int indexOf(Object o) {
+        return -1;
+    }
+
+    @Override
+    public int lastIndexOf(Object o) {
+        return -1;
+    }
+
+    @Override
+    public ListIterator<Object> listIterator() {
+        return Collections.emptyListIterator();
+    }
+
+    @Override
+    public ListIterator<Object> listIterator(int index) {
+        return Collections.emptyListIterator();
+    }
+
+    @Override
+    public List<Object> subList(int fromIndex, int toIndex) {
+        return NoOpList.get();
+    }
+}