--- a/aws-xray-recorder-sdk-core/src/main/java/com/amazonaws/xray/entities/NoOpSegment.java
+++ b/aws-xray-recorder-sdk-core/src/main/java/com/amazonaws/xray/entities/NoOpSegment.java
@@ -138,7 +138,7 @@ class NoOpSegment implements Segment {
     }
 
     @Override
-    public @Nullable String getNamespace() {
+    public String getNamespace() {
         return "";
     }