… aws-xray-recorder-sdk-apache-http/pom.xml
… aws-xray-recorder-sdk-apache-http/src/test/java/com/amazonaws/xray/proxies/apache/http/TracedHttpClientTest.java
… aws-xray-recorder-sdk-core/src/main/java/com/amazonaws/xray/entities/NoOpSet.java
… aws-xray-recorder-sdk-core/src/main/java/com/amazonaws/xray/entities/NoOpSubSegment.java
… aws-xray-recorder-sdk-core/src/main/java/com/amazonaws/xray/entities/Segment.java
… aws-xray-recorder-sdk-core/src/main/java/com/amazonaws/xray/entities/SegmentImpl.java
… aws-xray-recorder-sdk-core/src/main/java/com/amazonaws/xray/entities/Subsegment.java
… aws-xray-recorder-sdk-core/src/main/java/com/amazonaws/xray/entities/TraceID.java
… aws-xray-recorder-sdk-core/src/main/java/com/amazonaws/xray/javax/servlet/AWSXRayServletFilter.java
… aws-xray-recorder-sdk-core/src/test/java/com/amazonaws/xray/AWSXRayRecorderTest.java
… aws-xray-recorder-sdk-core/src/test/java/com/amazonaws/xray/javax/servlet/AWSXRayServletFilterTest.java
… config/checkstyle/checkstyle.xml