…o-op.

*Description of changes:*

- Rename `Dummy*` to `NoOp*` which is more inclusive
- Don't generate IDs for no-op segment / subsegment
- Create no-op subsegment when segment is not going to be emitted
- Reduce no-op implementation memory usage by using no-op singleton containers where possible

Notes
- I added `isRecording` only to solve one corner case where "forced sampling" is an emitted segment where `sampled == false`. I see forced sampling isn't implemented in the SDK so it'd have to be user code that uses it, since it exists I guess we want to support it, but ideally we can remove forced sampling (I don't see it in other SDKs). As far as I can tell, we have our own bugs related, this seems like it won't stream segments that were force sampled

https://github.com/aws/aws-xray-sdk-java/blob/master/aws-xray-recorder-sdk-core/src/main/java/com/amazonaws/xray/strategy/DefaultStreamingStrategy.java#L69

- I didn't add a name to the no-op, since they're not emitted I couldn't see a point for it

- I just used (0, 0) as a fixed invalid trace ID, I don't see any reason that would break anything downstream

- The no-op lock feels hacky, but is probably significant so I want to stick with it, in practice I'm not worried about breakage across JVM versions since JDK has to preserve compatibility because they didn't mark the class / methods final.

As we can expect, performance difference is quite large, 4-5x on these begin / end operations. I guess instrumentation currently operates on subsegments that won't be emitted, so such operations would also be optimized, and there should be a large reduction in memory usage by the SDK to boot.

After
```
AWSXRayRecorderBenchmark.beginDummySegmentBenchmark                                                               sample  515591     0.713 �   0.047   us/op
AWSXRayRecorderBenchmark.beginDummySegmentBenchmark:beginDummySegmentBenchmark�p0.00                              sample             0.400             us/op
AWSXRayRecorderBenchmark.beginDummySegmentBenchmark:beginDummySegmentBenchmark�p0.50                              sample             0.600             us/op
AWSXRayRecorderBenchmark.beginDummySegmentBenchmark:beginDummySegmentBenchmark�p0.90                              sample             0.800             us/op
AWSXRayRecorderBenchmark.beginDummySegmentBenchmark:beginDummySegmentBenchmark�p0.95                              sample             1.100             us/op
AWSXRayRecorderBenchmark.beginDummySegmentBenchmark:beginDummySegmentBenchmark�p0.99                              sample             1.400             us/op
AWSXRayRecorderBenchmark.beginDummySegmentBenchmark:beginDummySegmentBenchmark�p0.999                             sample            18.080             us/op
AWSXRayRecorderBenchmark.beginDummySegmentBenchmark:beginDummySegmentBenchmark�p0.9999                            sample            80.410             us/op
AWSXRayRecorderBenchmark.beginDummySegmentBenchmark:beginDummySegmentBenchmark�p1.00                              sample          2449.408             us/op
AWSXRayRecorderBenchmark.beginDummySegmentBenchmark:�gc.alloc.rate                                                sample      15  1128.141 � 133.568  MB/sec
AWSXRayRecorderBenchmark.beginDummySegmentBenchmark:�gc.alloc.rate.norm                                           sample      15  1632.436 �   0.054    B/op
AWSXRayRecorderBenchmark.beginDummySegmentBenchmark:�gc.churn.G1_Eden_Space                                       sample      15  1131.657 � 151.639  MB/sec
AWSXRayRecorderBenchmark.beginDummySegmentBenchmark:�gc.churn.G1_Eden_Space.norm                                  sample      15  1636.779 �  89.969    B/op
AWSXRayRecorderBenchmark.beginDummySegmentBenchmark:�gc.churn.G1_Survivor_Space                                   sample      15     0.037 �   0.025  MB/sec
AWSXRayRecorderBenchmark.beginDummySegmentBenchmark:�gc.churn.G1_Survivor_Space.norm                              sample      15     0.052 �   0.033    B/op
AWSXRayRecorderBenchmark.beginDummySegmentBenchmark:�gc.count                                                     sample      15   118.000            counts
AWSXRayRecorderBenchmark.beginDummySegmentBenchmark:�gc.time                                                      sample      15   207.000                ms
AWSXRayRecorderBenchmark.beginEndDummySegmentBenchmark                                                            sample  332321     1.570 �   0.082   us/op
AWSXRayRecorderBenchmark.beginEndDummySegmentBenchmark:beginEndDummySegmentBenchmark�p0.00                        sample             0.800             us/op
AWSXRayRecorderBenchmark.beginEndDummySegmentBenchmark:beginEndDummySegmentBenchmark�p0.50                        sample             1.200             us/op
AWSXRayRecorderBenchmark.beginEndDummySegmentBenchmark:beginEndDummySegmentBenchmark�p0.90                        sample             1.700             us/op
AWSXRayRecorderBenchmark.beginEndDummySegmentBenchmark:beginEndDummySegmentBenchmark�p0.95                        sample             2.400             us/op
AWSXRayRecorderBenchmark.beginEndDummySegmentBenchmark:beginEndDummySegmentBenchmark�p0.99                        sample             4.096             us/op
AWSXRayRecorderBenchmark.beginEndDummySegmentBenchmark:beginEndDummySegmentBenchmark�p0.999                       sample            32.737             us/op
AWSXRayRecorderBenchmark.beginEndDummySegmentBenchmark:beginEndDummySegmentBenchmark�p0.9999                      sample           181.924             us/op
AWSXRayRecorderBenchmark.beginEndDummySegmentBenchmark:beginEndDummySegmentBenchmark�p1.00                        sample          2412.544             us/op
AWSXRayRecorderBenchmark.beginEndDummySegmentBenchmark:�gc.alloc.rate                                             sample      15  1177.164 � 141.877  MB/sec
AWSXRayRecorderBenchmark.beginEndDummySegmentBenchmark:�gc.alloc.rate.norm                                        sample      15  2920.799 �   0.101    B/op
AWSXRayRecorderBenchmark.beginEndDummySegmentBenchmark:�gc.churn.G1_Eden_Space                                    sample      15  1176.485 � 165.956  MB/sec
AWSXRayRecorderBenchmark.beginEndDummySegmentBenchmark:�gc.churn.G1_Eden_Space.norm                               sample      15  2915.376 � 171.773    B/op
AWSXRayRecorderBenchmark.beginEndDummySegmentBenchmark:�gc.churn.G1_Survivor_Space                                sample      15     0.046 �   0.017  MB/sec
AWSXRayRecorderBenchmark.beginEndDummySegmentBenchmark:�gc.churn.G1_Survivor_Space.norm                           sample      15     0.113 �   0.040    B/op
AWSXRayRecorderBenchmark.beginEndDummySegmentBenchmark:�gc.count                                                  sample      15   123.000            counts
AWSXRayRecorderBenchmark.beginEndDummySegmentBenchmark:�gc.time                                                   sample      15   218.000                ms
AWSXRayRecorderBenchmark.beginEndDummySegmentSubsegmentBenchmark                                                  sample  386255     2.382 �   0.084   us/op
AWSXRayRecorderBenchmark.beginEndDummySegmentSubsegmentBenchmark:beginEndDummySegmentSubsegmentBenchmark�p0.00    sample             1.300             us/op
AWSXRayRecorderBenchmark.beginEndDummySegmentSubsegmentBenchmark:beginEndDummySegmentSubsegmentBenchmark�p0.50    sample             1.800             us/op
AWSXRayRecorderBenchmark.beginEndDummySegmentSubsegmentBenchmark:beginEndDummySegmentSubsegmentBenchmark�p0.90    sample             3.200             us/op
AWSXRayRecorderBenchmark.beginEndDummySegmentSubsegmentBenchmark:beginEndDummySegmentSubsegmentBenchmark�p0.95    sample             4.000             us/op
AWSXRayRecorderBenchmark.beginEndDummySegmentSubsegmentBenchmark:beginEndDummySegmentSubsegmentBenchmark�p0.99    sample             6.000             us/op
AWSXRayRecorderBenchmark.beginEndDummySegmentSubsegmentBenchmark:beginEndDummySegmentSubsegmentBenchmark�p0.999   sample            41.055             us/op
AWSXRayRecorderBenchmark.beginEndDummySegmentSubsegmentBenchmark:beginEndDummySegmentSubsegmentBenchmark�p0.9999  sample           191.166             us/op
AWSXRayRecorderBenchmark.beginEndDummySegmentSubsegmentBenchmark:beginEndDummySegmentSubsegmentBenchmark�p1.00    sample          2666.496             us/op
AWSXRayRecorderBenchmark.beginEndDummySegmentSubsegmentBenchmark:�gc.alloc.rate                                   sample      15  1159.502 � 138.181  MB/sec
AWSXRayRecorderBenchmark.beginEndDummySegmentSubsegmentBenchmark:�gc.alloc.rate.norm                              sample      15  4489.190 �   0.136    B/op
AWSXRayRecorderBenchmark.beginEndDummySegmentSubsegmentBenchmark:�gc.churn.G1_Eden_Space                          sample      15  1154.082 � 150.592  MB/sec
AWSXRayRecorderBenchmark.beginEndDummySegmentSubsegmentBenchmark:�gc.churn.G1_Eden_Space.norm                     sample      15  4471.361 � 304.775    B/op
AWSXRayRecorderBenchmark.beginEndDummySegmentSubsegmentBenchmark:�gc.churn.G1_Survivor_Space                      sample      15     0.045 �   0.026  MB/sec
AWSXRayRecorderBenchmark.beginEndDummySegmentSubsegmentBenchmark:�gc.churn.G1_Survivor_Space.norm                 sample      15     0.170 �   0.091    B/op
AWSXRayRecorderBenchmark.beginEndDummySegmentSubsegmentBenchmark:�gc.count                                        sample      15   122.000            counts
AWSXRayRecorderBenchmark.beginEndDummySegmentSubsegmentBenchmark:�gc.time                                         sample      15   215.000                ms
AWSXRayRecorderBenchmark.beginSubsegmentDummyParentBenchmark                                                      sample  417494     0.420 �   0.040   us/op
AWSXRayRecorderBenchmark.beginSubsegmentDummyParentBenchmark:beginSubsegmentDummyParentBenchmark�p0.00            sample             0.200             us/op
AWSXRayRecorderBenchmark.beginSubsegmentDummyParentBenchmark:beginSubsegmentDummyParentBenchmark�p0.50            sample             0.300             us/op
AWSXRayRecorderBenchmark.beginSubsegmentDummyParentBenchmark:beginSubsegmentDummyParentBenchmark�p0.90            sample             0.500             us/op
AWSXRayRecorderBenchmark.beginSubsegmentDummyParentBenchmark:beginSubsegmentDummyParentBenchmark�p0.95            sample             0.600             us/op
AWSXRayRecorderBenchmark.beginSubsegmentDummyParentBenchmark:beginSubsegmentDummyParentBenchmark�p0.99            sample             0.900             us/op
AWSXRayRecorderBenchmark.beginSubsegmentDummyParentBenchmark:beginSubsegmentDummyParentBenchmark�p0.999           sample            14.192             us/op
AWSXRayRecorderBenchmark.beginSubsegmentDummyParentBenchmark:beginSubsegmentDummyParentBenchmark�p0.9999          sample            46.033             us/op
AWSXRayRecorderBenchmark.beginSubsegmentDummyParentBenchmark:beginSubsegmentDummyParentBenchmark�p1.00            sample          2560.000             us/op
AWSXRayRecorderBenchmark.beginSubsegmentDummyParentBenchmark:�gc.alloc.rate                                       sample      15  1171.686 � 107.395  MB/sec
AWSXRayRecorderBenchmark.beginSubsegmentDummyParentBenchmark:�gc.alloc.rate.norm                                  sample      15  3008.739 �   0.076    B/op
AWSXRayRecorderBenchmark.beginSubsegmentDummyParentBenchmark:�gc.churn.G1_Eden_Space                              sample      15  1177.227 � 131.248  MB/sec
AWSXRayRecorderBenchmark.beginSubsegmentDummyParentBenchmark:�gc.churn.G1_Eden_Space.norm                         sample      15  3021.447 � 155.657    B/op
AWSXRayRecorderBenchmark.beginSubsegmentDummyParentBenchmark:�gc.churn.G1_Survivor_Space                          sample      15     0.028 �   0.013  MB/sec
AWSXRayRecorderBenchmark.beginSubsegmentDummyParentBenchmark:�gc.churn.G1_Survivor_Space.norm                     sample      15     0.070 �   0.033    B/op
AWSXRayRecorderBenchmark.beginSubsegmentDummyParentBenchmark:�gc.count                                            sample      15   123.000            counts
AWSXRayRecorderBenchmark.beginSubsegmentDummyParentBenchmark:�gc.time                                             sample      15   213.000                ms
AWSXRayRecorderBenchmark.endDummySegmentBenchmark                                                                 sample  473975     0.858 �   0.045   us/op
AWSXRayRecorderBenchmark.endDummySegmentBenchmark:endDummySegmentBenchmark�p0.00                                  sample             0.400             us/op
AWSXRayRecorderBenchmark.endDummySegmentBenchmark:endDummySegmentBenchmark�p0.50                                  sample             0.700             us/op
AWSXRayRecorderBenchmark.endDummySegmentBenchmark:endDummySegmentBenchmark�p0.90                                  sample             1.000             us/op
AWSXRayRecorderBenchmark.endDummySegmentBenchmark:endDummySegmentBenchmark�p0.95                                  sample             1.300             us/op
AWSXRayRecorderBenchmark.endDummySegmentBenchmark:endDummySegmentBenchmark�p0.99                                  sample             2.000             us/op
AWSXRayRecorderBenchmark.endDummySegmentBenchmark:endDummySegmentBenchmark�p0.999                                 sample            18.400             us/op
AWSXRayRecorderBenchmark.endDummySegmentBenchmark:endDummySegmentBenchmark�p0.9999                                sample           117.558             us/op
AWSXRayRecorderBenchmark.endDummySegmentBenchmark:endDummySegmentBenchmark�p1.00                                  sample          2904.064             us/op
AWSXRayRecorderBenchmark.endDummySegmentBenchmark:�gc.alloc.rate                                                  sample      15  1144.357 � 148.093  MB/sec
AWSXRayRecorderBenchmark.endDummySegmentBenchmark:�gc.alloc.rate.norm                                             sample      15  3608.961 �   0.192    B/op
AWSXRayRecorderBenchmark.endDummySegmentBenchmark:�gc.churn.G1_Eden_Space                                         sample      15  1139.353 � 181.437  MB/sec
AWSXRayRecorderBenchmark.endDummySegmentBenchmark:�gc.churn.G1_Eden_Space.norm                                    sample      15  3586.506 � 239.624    B/op
AWSXRayRecorderBenchmark.endDummySegmentBenchmark:�gc.churn.G1_Survivor_Space                                     sample      15     0.052 �   0.045  MB/sec
AWSXRayRecorderBenchmark.endDummySegmentBenchmark:�gc.churn.G1_Survivor_Space.norm                                sample      15     0.161 �   0.135    B/op
AWSXRayRecorderBenchmark.endDummySegmentBenchmark:�gc.count                                                       sample      15   112.000            counts
AWSXRayRecorderBenchmark.endDummySegmentBenchmark:�gc.time                                                        sample      15   202.000                ms
AWSXRayRecorderBenchmark.endSubsegmentDummyParentBenchmark                                                        sample  468450     0.487 �   0.022   us/op
AWSXRayRecorderBenchmark.endSubsegmentDummyParentBenchmark:endSubsegmentDummyParentBenchmark�p0.00                sample             0.200             us/op
AWSXRayRecorderBenchmark.endSubsegmentDummyParentBenchmark:endSubsegmentDummyParentBenchmark�p0.50                sample             0.400             us/op
AWSXRayRecorderBenchmark.endSubsegmentDummyParentBenchmark:endSubsegmentDummyParentBenchmark�p0.90                sample             0.600             us/op
AWSXRayRecorderBenchmark.endSubsegmentDummyParentBenchmark:endSubsegmentDummyParentBenchmark�p0.95                sample             0.800             us/op
AWSXRayRecorderBenchmark.endSubsegmentDummyParentBenchmark:endSubsegmentDummyParentBenchmark�p0.99                sample             1.100             us/op
AWSXRayRecorderBenchmark.endSubsegmentDummyParentBenchmark:endSubsegmentDummyParentBenchmark�p0.999               sample            14.645             us/op
AWSXRayRecorderBenchmark.endSubsegmentDummyParentBenchmark:endSubsegmentDummyParentBenchmark�p0.9999              sample            51.949             us/op
AWSXRayRecorderBenchmark.endSubsegmentDummyParentBenchmark:endSubsegmentDummyParentBenchmark�p1.00                sample          2486.272             us/op
AWSXRayRecorderBenchmark.endSubsegmentDummyParentBenchmark:�gc.alloc.rate                                         sample      15  1174.275 � 129.432  MB/sec
AWSXRayRecorderBenchmark.endSubsegmentDummyParentBenchmark:�gc.alloc.rate.norm                                    sample      15  3200.793 �   0.107    B/op
AWSXRayRecorderBenchmark.endSubsegmentDummyParentBenchmark:�gc.churn.G1_Eden_Space                                sample      15  1168.499 � 139.894  MB/sec
AWSXRayRecorderBenchmark.endSubsegmentDummyParentBenchmark:�gc.churn.G1_Eden_Space.norm                           sample      15  3189.344 � 239.095    B/op
AWSXRayRecorderBenchmark.endSubsegmentDummyParentBenchmark:�gc.churn.G1_Survivor_Space                            sample      15     0.028 �   0.014  MB/sec
AWSXRayRecorderBenchmark.endSubsegmentDummyParentBenchmark:�gc.churn.G1_Survivor_Space.norm                       sample      15     0.074 �   0.037    B/op
AWSXRayRecorderBenchmark.endSubsegmentDummyParentBenchmark:�gc.count                                              sample      15   122.000            counts
AWSXRayRecorderBenchmark.endSubsegmentDummyParentBenchmark:�gc.time                                               sample      15   216.000                ms
```

Before
```
Benchmark                                                                                                           Mode     Cnt     Score     Error   Units
AWSXRayRecorderBenchmark.beginDummySegmentBenchmark                                                               sample  372444     2.403 �   0.089   us/op
AWSXRayRecorderBenchmark.beginDummySegmentBenchmark:beginDummySegmentBenchmark�p0.00                              sample             1.500             us/op
AWSXRayRecorderBenchmark.beginDummySegmentBenchmark:beginDummySegmentBenchmark�p0.50                              sample             2.000             us/op
AWSXRayRecorderBenchmark.beginDummySegmentBenchmark:beginDummySegmentBenchmark�p0.90                              sample             2.300             us/op
AWSXRayRecorderBenchmark.beginDummySegmentBenchmark:beginDummySegmentBenchmark�p0.95                              sample             3.400             us/op
AWSXRayRecorderBenchmark.beginDummySegmentBenchmark:beginDummySegmentBenchmark�p0.99                              sample             5.696             us/op
AWSXRayRecorderBenchmark.beginDummySegmentBenchmark:beginDummySegmentBenchmark�p0.999                             sample            35.840             us/op
AWSXRayRecorderBenchmark.beginDummySegmentBenchmark:beginDummySegmentBenchmark�p0.9999                            sample           172.481             us/op
AWSXRayRecorderBenchmark.beginDummySegmentBenchmark:beginDummySegmentBenchmark�p1.00                              sample          2326.528             us/op
AWSXRayRecorderBenchmark.beginDummySegmentBenchmark:�gc.alloc.rate                                                sample      15   574.153 �  25.315  MB/sec
AWSXRayRecorderBenchmark.beginDummySegmentBenchmark:�gc.alloc.rate.norm                                           sample      15  2305.089 �  12.529    B/op
AWSXRayRecorderBenchmark.beginDummySegmentBenchmark:�gc.churn.G1_Eden_Space                                       sample      15   577.312 �  60.074  MB/sec
…