@anuraaga 2020-07-16
Yup sorry for missing replying to your comment about the lock - yeah I think the listener should work ok in general. If we find there's a different use case we can add safer APIs instead, straightforward is accepting `Consumer<List<Subsegment>>`